requirements.txt
# --- Core Framework ---
langchain-core
langchain-groq
langgraph
pydantic

# --- Research & Data ---
tavily-python
beautifulsoup4
requests
pandas
pypdf

# --- Local LLM & Utilities ---
langchain-ollama
python-dotenv
pyyaml
